Modify WinPort, VT, dialog.cpp (from far2l)

+++ WinPort/src/Backend/Backend.h
@@ -46,7 +46,7 @@ public:
 	virtual void OnConsoleSetCursorBlinkTime(DWORD interval) = 0;
 
 	virtual void OnGetConsoleImageCaps(WinportGraphicsInfo *wgi) = 0;
-	virtual bool OnSetConsoleImage(const char *id, DWORD flags, COORD pos, DWORD width, DWORD height, const void *buffer) = 0;
+	virtual bool OnSetConsoleImage(const char *id, DWORD64 flags, COORD pos, DWORD width, DWORD height, const void *buffer) = 0;
 	virtual bool OnDeleteConsoleImage(const char *id) = 0;
 
 	virtual const char *OnConsoleBackendInfo(int entity) = 0;
@@ -224,7 +224,7 @@ public:
 	virtual void RepaintsDeferFinish(bool force) = 0;
 
 	virtual void OnGetConsoleImageCaps(WinportGraphicsInfo *wgi) = 0;
-	virtual bool OnSetConsoleImage(const char *id, DWORD flags, COORD pos, DWORD width, DWORD height, const void *buffer) = 0;
+	virtual bool OnSetConsoleImage(const char *id, DWORD64 flags, COORD pos, DWORD width, DWORD height, const void *buffer) = 0;
 	virtual bool OnDeleteConsoleImage(const char *id) = 0;
 
 	virtual const char *BackendInfo(int entity) = 0;
