+++ WinPort/src/Backend/WX/wxMain.h
@@ -180,7 +180,7 @@ class WinPortPanel: public wxPanel, protected IConsoleOutputBackend
 	virtual void OnConsoleSetCursorBlinkTime(DWORD interval);
 	virtual const char *OnConsoleBackendInfo(int entity);
 	virtual void OnGetConsoleImageCaps(WinportGraphicsInfo *wgi);
-	virtual bool OnSetConsoleImage(const char *id, DWORD flags, COORD pos, DWORD width, DWORD height, const void *buffer);
+	virtual bool OnSetConsoleImage(const char *id, DWORD64 flags, COORD pos, DWORD width, DWORD height, const void *buffer);
 	virtual bool OnDeleteConsoleImage(const char *id);
 
 public:
